+++ packages/component/src/components/ButtonGroup/button-group.tsx
@@ -3,6 +3,7 @@ import clsx from 'clsx'
 import { splitProps } from 'solid-js'
 import { useConfig } from '../ConfigProvider'
 
+/** 按钮组 */
 export function ButtonGroup(props: JSX.HTMLAttributes<HTMLDivElement>) {
   const [local, others] = splitProps(props, ['class'])
   const config = useConfig()