+++ packages/component/src/components/Dialog/__demo__/basic.tsx
@@ -1,0 +1,12 @@
+import { Dialog } from '@thinke/idle-ui'
+
+export default () => {
+  return (
+    <div class="">
+      <Dialog>
+        <p>对话框 Dialog显示在主要内容上方的模态窗口 1</p>
+        <div class="w-200 h-20 bg-blue"></div>
+      </Dialog>
+    </div>
+  )
+}