+++ packages/component/src/components/Dialog/style/index.ts
@@ -1,0 +1,3 @@
+import '../../_style/theme.less'
+
+import './dialog.less'