+++ packages/component/src/utils/sleep.ts
@@ -1,0 +1,5 @@
+/**
+ * 休眠一定时间
+ * @param ms 休眠时间，单位ms @default 0
+ */
+export const sleep = (ms: number) => new Promise(resolve => setTimeout(resolve, ms))