+++ packages/component/CHANGELOG.md
@@ -1,3 +1,18 @@
+## <small>0.0.3 (2025-11-04)</small>
+
+
+
+
+## <small>0.0.2 (2025-11-04)</small>
+
+* ✨ feat(Button): text hover 样式调整 ([fa3a421](https://github.com/thinke5/idle-ui/commit/fa3a421))
+* 🎈 perf(doc): 按需引入css、代码块优化 ([d96759c](https://github.com/thinke5/idle-ui/commit/d96759c))
+* 🦄 refactor(ButtonGroup): 调整代码位置 ([ec4a5f7](https://github.com/thinke5/idle-ui/commit/ec4a5f7))
+* 🧪 test: 增加demo测试 ([cab1294](https://github.com/thinke5/idle-ui/commit/cab1294))
+* 🧪 test(demo): 按需加载测试 ([37f7ab5](https://github.com/thinke5/idle-ui/commit/37f7ab5))
+
+
+
 ## <small>0.0.1 (2025-10-28)</small>
 
 * ✨ feat(button): 初版 ([5e7c14f](https://github.com/thinke5/idle-ui/commit/5e7c14f))